pick up concurrency API 1.0-b05 and RI 1.0-b04
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@60346 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ appserver/concurrent/runtime/src/test/java/org/glassfish/concurrent/runtime/ConcurrentRuntimeTest.java
@@ -175,7 +175,7 @@ public class ConcurrentRuntimeTest {
         final int CORE_POOL_SIZE = 1;
         final int MAXIMUM_POOL_SIZE = 5;
         final boolean LONG_RUNNING_TASKS = true;
-        final int KEEP_ALIVE_SECONDS = 88;
+        final long KEEP_ALIVE_SECONDS = 88L;
         final int TASK_QUEUE_CAPACITY = 12345;
